+++ src/advisories/helpers.rs
@@ -270,201 +270,6 @@ impl Report {
             .chain(self.unmaintained.iter().map(|wi| (Kind::Unmaintained, wi)))
             .chain(self.unsound.iter().map(|wi| (Kind::Unsound, wi)))
     }
-
-    pub fn gather_patches(&self, krates: &Krates, force: bool) -> Result<PatchSet<'_>, Error> {
-        use krates::petgraph as pg;
-        use pg::{visit::EdgeRef, Direction};
-
-        let graph = krates.graph();
-        let index = Index::load(krates)?;
-
-        let mut patches = Vec::new();
-        let mut diags = Vec::new();
-
-        for vuln in &self.vulnerabilities {
-            // 1. Get the package with the vulnerability
-            // 2. Recursively walk up the dependency chain until we've reach all roots
-            // (workspace crates) that depend on the vulnerable crate version
-            // 3. For each crate in the chain, check to see if has a version
-            // available that ultimately includes a patched version of the vulnerable crate
-            let (ind, vuln_krate) = krate_for_pkg(krates, &vuln.package).unwrap();
-
-            // We could also see if there are unaffected versions, but easier to just say we only support
-            // fixing by using newer versions than older versions
-            if vuln.versions.patched.is_empty() {
-                // let mut pack = Pack::with_kid(Check::Advisories, vuln_krate.id.clone());
-                // pack.push(
-                //     Diagnostic::error()
-                //         .with_message()
-                //         .with_labels(vec![ctx.label_for_span(i.index(), message)])
-                //         .with_code(id.as_str().to_owned())
-                //         .with_notes(notes),
-                // );
-                continue;
-            }
-
-            // Gather the versions that fix the issue. Note that there could be cases where the fix is in disjoint
-            // versions.
-            let required: Vec<_> = {
-                let index_krate = match index.read_krate(vuln_krate) {
-                    Some(ik) => ik,
-                    None => {
-                        continue;
-                    }
-                };
-
-                let patched = &vuln.versions.patched;
-
-                index_krate
-                    .versions()
-                    .iter()
-                    .filter_map(|vs| {
-                        vs.version()
-                            .parse()
-                            .ok()
-                            .and_then(|version: rustsec::Version| {
-                                if patched.iter().any(|v| v.matches(&version)) {
-                                    // rustsec uses an older version of semver
-                                    vs.version().parse::<Version>().ok()
-                                } else {
-                                    None
-                                }
-                            })
-                    })
-                    .collect()
-            };
-
-            if required.is_empty() {
-                // let mut pack = Pack::with_kid(Check::Advisories, vuln_krate.id.clone());
-                // pack.push(
-                //     Diagnostic::error()
-                //         .with_message()
-                //         .with_labels(vec![ctx.label_for_span(i.index(), message)])
-                //         .with_code(id.as_str().to_owned())
-                //         .with_notes(notes),
-                // );
-                continue;
-            }
-
-            let mut krate_stack = vec![(ind, vuln_krate, required)];
-            let mut visited = std::collections::HashSet::new();
-
-            while let Some((nid, krate, available)) = krate_stack.pop() {
-                for edge in graph.edges_directed(nid, Direction::Incoming) {
-                    // We only need to visit each unique edge once
-                    if !visited.insert(edge.id()) {
-                        continue;
-                    }
-
-                    let parent_id = edge.source();
-                    let parent = &graph[parent_id];
-
-                    if let Some(src) = &parent.krate.source {
-                        if src.is_registry() {
-                            match Self::find_possible_versions(&index, parent, krate, required) {
-                                Ok(required) => {}
-                                Err(e) => {
-                                    continue;
-                                }
-                            }
-                        } else if !src.is_path() {
-                            // TODO: Emit warning that we can't update
-                            continue;
-                        }
-                    }
-
-                    patches.push(Patch {
-                        vuln,
-                        vuln_krate,
-                        manifest_to_patch: &parent.krate,
-                        crate_to_patch: (),
-                    });
-                }
-            }
-        }
-
-        Ok(PatchSet {
-            patches,
-            diagnostics: diags,
-        })
-    }
-
-    fn find_possible_versions(
-        index: &Index,
-        parent: &Krate,
-        child: &Krate,
-        required: &[Version],
-    ) -> Result<Vec<Version>, Error> {
-        // Lookup the available versions in the registry's index to see if there is a
-        // compatible version we can update to
-        let parent_krate = index.read_krate(&parent).with_context(|| {
-            format!(
-                "Unable to find registry index entry for crate '{}'",
-                parent.name
-            )
-        })?;
-
-        // Grab all of the versions of the parent crate that have a version requirement that accepts
-        // any of the specified required versions
-        let available: Vec<_> = parent_krate.versions().iter().filter_map(|kv| {
-            let version: Version = match kv.version().parse() {
-                Ok(vs) => {
-                    if vs < parent.version {
-                        return None;
-                    }
-
-                    vs
-                },
-                Err(err) => {
-                    log::warn!("Unable to parse version '{}' for index crate '{}': {}", kv.version(), parent.name, err);
-                    return None;
-                }
-            };
-
-
-            if let Some(dep) = kv.dependencies().iter().find(|dep| dep.kind() != crates_index::DependencyKind::Dev && dep.name() == child.name) {
-                let req: VersionReq = match dep.requirement().parse() {
-                    Ok(req) => req,
-                    Err(err) => {
-                        log::warn!("Unable to parse version requirement '{}' for index crate '{}', dependency '{}': {}", dep.requirement(), parent.name, child.name, err);
-                        return None;
-                    }
-                };
-
-                if !required.iter().any(|vs| req.matches(&vs)) {
-                    return None;
-                }
-            }
-
-            // If the dependency has been removed from a future version, it is
-            // also a candidate
-            Some(version)
-        }).collect();
-
-        if available.is_empty() {
-            anyhow::bail!("No versions were available that used a required version of the crate");
-        } else {
-            Ok(available)
-        }
-    }
-}
-
-//enum PatchLink<'a> {}
-
-pub struct Patch<'a> {
-    /// The vulnerability the patch is attempting to address
-    pub vuln: &'a Vulnerability,
-    /// The vulnerable crate
-    pub vuln_krate: &'a Krate,
-    /// The crate manifest we want to patch to update versions
-    pub manifest_to_patch: &'a Krate,
-    /// The crate and version we want to update to get the fix
-    pub crate_to_patch: (&'a str, Version),
-}
-
-pub struct PatchSet<'a> {
-    diagnostics: Vec<Pack>,
-    patches: Vec<Patch<'a>>,
 }
 
 #[cfg(test)]
