+++ src/advisories/mod.rs
@@ -1,4 +1,6 @@
 pub mod cfg;
+#[cfg(feature = "fix")]
+mod fix;
 mod helpers;
 
 use crate::{
